Split memory map routines into their own file
